+++ server/scripts/check_auth.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 
 # DeepFellow Software Framework.
-# Copyright © 2025 Simplito sp. z o.o.
+# Copyright © 2026 Simplito sp. z o.o.
 #
 # This file is part of the DeepFellow Software Framework (https://deepfellow.ai).
 # This software is Licensed under the DeepFellow Free License.